+++ tests/Remove-GTUserEntitlements.Tests.ps1
@@ -1,7 +1,28 @@
-. "$PSScriptRoot/../functions/Remove-GTUserEntitlements.ps1"
-
 Describe "Remove-GTUserEntitlements" {
     BeforeAll {
+        # Create stub functions for all cmdlets used (before sourcing the function)
+        function Get-MgContext { }
+        function Install-GTRequiredModule { }
+        function Get-MgBetaUser { }
+        function Remove-GTUserGroupMemberships { }
+        function Remove-GTUserGroupOwnerships { }
+        function Remove-GTUserLicenses { }
+        function Remove-GTUserServicePrincipalOwnerships { }
+        function Remove-GTUserEnterpriseAppOwnership { }
+        function Remove-GTUserAppRoleAssignments { }
+        function Remove-GTUserRoleAssignments { }
+        function Remove-GTPIMRoleEligibility { }
+        function Remove-GTUserAdministrativeUnitMemberships { }
+        function Remove-GTUserAccessPackageAssignments { }
+        function Remove-GTUserDelegatedPermissionGrants { }
+        function Write-PSFMessage { }
+        
+        # Source the GTValidation script for UPN regex
+        . "$PSScriptRoot/../internal/functions/GTValidation.ps1"
+        
+        # Now source the function under test
+        . "$PSScriptRoot/../functions/Remove-GTUserEntitlements.ps1"
+        
         # Mock the required modules and functions
         Mock -CommandName "Get-MgContext" -MockWith { 
             [PSCustomObject]@{
@@ -22,6 +43,7 @@ Describe "Remove-GTUserEntitlements" {
         Mock -CommandName "Remove-GTUserEnterpriseAppOwnership" -MockWith { }
         Mock -CommandName "Remove-GTUserAppRoleAssignments" -MockWith { }
         Mock -CommandName "Remove-GTUserRoleAssignments" -MockWith { }
+        Mock -CommandName "Remove-GTPIMRoleEligibility" -MockWith { }
         Mock -CommandName "Remove-GTUserAdministrativeUnitMemberships" -MockWith { }
         Mock -CommandName "Remove-GTUserAccessPackageAssignments" -MockWith { }
         Mock -CommandName "Remove-GTUserDelegatedPermissionGrants" -MockWith { }
@@ -41,6 +63,11 @@ Describe "Remove-GTUserEntitlements" {
         }
 
         It "should accept valid UPN format" {
+            Mock -CommandName "Get-MgContext" -MockWith { 
+                [PSCustomObject]@{
+                    Scopes = @('GroupMember.ReadWrite.All', 'Group.ReadWrite.All', 'Directory.ReadWrite.All', 'RoleManagement.ReadWrite.Directory', 'RoleEligibilitySchedule.ReadWrite.Directory', 'AdministrativeUnit.ReadWrite.All', 'EntitlementManagement.ReadWrite.All', 'DelegatedPermissionGrant.ReadWrite.All')
+                }
+            }
             { Remove-GTUserEntitlements -UserUPNs "test@contoso.com" -removeAll -WhatIf } | Should -Not -Throw
         }
     }
@@ -53,6 +80,56 @@ Describe "Remove-GTUserEntitlements" {
                 }
             }
             { Remove-GTUserEntitlements -UserUPNs "test@contoso.com" -removeAll } | Should -Throw "*Required scopes are missing*"
+        }
+
+        It "should include RoleEligibilitySchedule.ReadWrite.Directory in required scopes" {
+            Mock -CommandName "Get-MgContext" -MockWith { 
+                [PSCustomObject]@{
+                    Scopes = @('GroupMember.ReadWrite.All', 'Group.ReadWrite.All', 'Directory.ReadWrite.All', 'RoleManagement.ReadWrite.Directory', 'AdministrativeUnit.ReadWrite.All', 'EntitlementManagement.ReadWrite.All', 'DelegatedPermissionGrant.ReadWrite.All')
+                }
+            }
+            { Remove-GTUserEntitlements -UserUPNs "test@contoso.com" -removeAll } | Should -Throw "*Required scopes are missing*RoleEligibilitySchedule.ReadWrite.Directory*"
+        }
+    }
+
+    Context "PIM Role Eligibility Removal" {
+        It "should call Remove-GTPIMRoleEligibility when removePIMRoleEligibility is specified" {
+            Mock -CommandName "Get-MgContext" -MockWith { 
+                [PSCustomObject]@{
+                    Scopes = @('GroupMember.ReadWrite.All', 'Group.ReadWrite.All', 'Directory.ReadWrite.All', 'RoleManagement.ReadWrite.Directory', 'RoleEligibilitySchedule.ReadWrite.Directory', 'AdministrativeUnit.ReadWrite.All', 'EntitlementManagement.ReadWrite.All', 'DelegatedPermissionGrant.ReadWrite.All')
+                }
+            }
+            Mock -CommandName "Remove-GTPIMRoleEligibility" -MockWith { }
+            
+            Remove-GTUserEntitlements -UserUPNs "test@contoso.com" -removePIMRoleEligibility -WhatIf
+            
+            Should -Invoke -CommandName "Remove-GTPIMRoleEligibility" -Times 1
+        }
+
+        It "should call Remove-GTPIMRoleEligibility when removeAll is specified" {
+            Mock -CommandName "Get-MgContext" -MockWith { 
+                [PSCustomObject]@{
+                    Scopes = @('GroupMember.ReadWrite.All', 'Group.ReadWrite.All', 'Directory.ReadWrite.All', 'RoleManagement.ReadWrite.Directory', 'RoleEligibilitySchedule.ReadWrite.Directory', 'AdministrativeUnit.ReadWrite.All', 'EntitlementManagement.ReadWrite.All', 'DelegatedPermissionGrant.ReadWrite.All')
+                }
+            }
+            Mock -CommandName "Remove-GTPIMRoleEligibility" -MockWith { }
+            
+            Remove-GTUserEntitlements -UserUPNs "test@contoso.com" -removeAll -WhatIf
+            
+            Should -Invoke -CommandName "Remove-GTPIMRoleEligibility" -Times 1
+        }
+
+        It "should not call Remove-GTPIMRoleEligibility when removePIMRoleEligibility is not specified" {
+            Mock -CommandName "Get-MgContext" -MockWith { 
+                [PSCustomObject]@{
+                    Scopes = @('GroupMember.ReadWrite.All', 'Group.ReadWrite.All', 'Directory.ReadWrite.All', 'RoleManagement.ReadWrite.Directory', 'RoleEligibilitySchedule.ReadWrite.Directory', 'AdministrativeUnit.ReadWrite.All', 'EntitlementManagement.ReadWrite.All', 'DelegatedPermissionGrant.ReadWrite.All')
+                }
+            }
+            Mock -CommandName "Remove-GTPIMRoleEligibility" -MockWith { }
+            
+            Remove-GTUserEntitlements -UserUPNs "test@contoso.com" -removeGroups -WhatIf
+            
+            Should -Invoke -CommandName "Remove-GTPIMRoleEligibility" -Times 0
         }
     }
 }